Refactor TranslatorGemini, moved it to the GeminiTranslatorService
