+++ frontend/src/pages/Settings/Subtitles/index.tsx
@@ -554,6 +554,7 @@ const SettingsSubtitlesView: FunctionComponent = () => {
             label="Lingarr endpoint"
             settingKey="settings-translator-lingarr_url"
           />
+          <Message>Base URL of Lingarr (e.g., http://localhost:9876)</Message>
         </CollapseBox>
         <Check
           label="Add translation info at the beginning"
